Controls actually work now
Turns out, you actually DO need to invert motors, and no, the lights on the controllers don't communicate that at all. Anyway, it works now (copy-pasted stuff from OI, Robot, and RobotContainer in robot2021 to make it work better). Anyway, we now have actual working teleop code (it's super fast and has no deadzones, so that's not *great*, but it's fine for now). Anyway, uh, good?